reorganise form stories to fit with the vulcan-ui-material package
